[21-06-11] 커밋

+++ src/main/webapp/WEB-INF/views/board/content.jsp
@@ -9,12 +9,31 @@
    <title>Insert title here</title>
 
    <%@ include file="../include/static-head.jsp" %>
+   <style>
+      .attach-file-list {
+         display: flex;
+      }
 
+      .attach-file-list a {
+         display: flex;
+         flex-direction: column;
+      }
+
+      .attach-file-list a img {
+         width: 100px;
+         height: 100px;
+         display: block;
+      }
+
+      .attach-file-list .thumbnail-box {
+         display: flex;
+      }
+   </style>
 </head>
 
 <body>
-   <%@ include file="../include/header.jsp" %>
 
+   <%@ include file="../include/header.jsp" %>
    <div class="container">
       <div class="row">
          <div class="offset-md-1 col-md-10">
@@ -34,7 +53,13 @@
             <c:if test="${article.writer == loginUser.account || loginUser.auth == 'ADMIN'}">
                <a href="/board/modify?boardNo=${article.boardNo}&vf=false">글 수정하기</a>
             </c:if>
+
          </div>
+      </div>
+
+      <!-- 첨부파일 영역 -->
+      <div class="row">
+         <div class="attach-file-list"></div>
       </div>
 
       <!-- 댓글 영역 -->
@@ -75,15 +100,15 @@
                <div id="replyCollapse" class="card">
                   <div id="replyData">
                      <!-- 
-								< JS로 댓글 정보 DIV삽입 > 
-							-->
+                        < JS로 댓글 정보 DIV삽입 > 
+                     -->
                   </div>
 
                   <!-- 댓글 페이징 영역 -->
                   <ul class="pagination justify-content-center">
                      <!-- 
-								< JS로 댓글 페이징 DIV삽입 > 
-							-->
+                        < JS로 댓글 페이징 DIV삽입 > 
+                     -->
                   </ul>
                </div>
             </div> <!-- end reply content -->
@@ -99,7 +124,7 @@
             <!-- Modal Header -->
             <div class="modal-header" style="background: #343A40; color: white;">
                <h4 class="modal-title">댓글 수정하기</h4>
-               <button type="button" class="close text-white" data-dismiss="modal">&times;</button>
+               <button type="button" class="close text-white" data-dismiss="modal">X</button>
             </div>
 
             <!-- Modal body -->
@@ -107,7 +132,8 @@
                <div class="form-group">
                   <input id="modReplyId" type="hidden">
                   <label for="modReplyText" hidden>댓글내용</label>
-                  <textarea id="modReplyText" class="form-control" placeholder="수정할 댓글 내용을 입력하세요." rows="3"></textarea>
+                  <textarea id="modReplyText" class="form-control" placeholder="수정할 댓글 내용을 입력하세요."
+                     rows="3"></textarea>
                </div>
             </div>
 
@@ -127,11 +153,81 @@
 
    <%@ include file="../include/footer.jsp" %>
 
+
+   <!-- 첨부파일 관련 스크립트 -->
+   <script>
+      $(function () {
+         const boardNo = '${article.boardNo}';
+         const $attachDiv = $('.attach-file-list');
+
+         //첨부파일 경로 목록 요청
+         fetch('/board/file/' + boardNo)
+            .then(res => res.json())
+            .then(filePathList => {
+               showFileData(filePathList)
+            });
+
+
+         //드롭한 파일의 형식에 따라 태그를 만들어주는 함수
+         function showFileData(pathList) {
+            //경로: \2021\06\08\dfjskfdjskf_dfjskfdj_dog.gif
+            for (let path of pathList) {
+               //이미지인지 아닌지에 따라 구분하여 처리
+               checkExtType(path);
+            }
+         }
+
+         //확장자 판별 후 태그 생성 처리 함수
+         function checkExtType(path) {
+            //원본 파일명 추출
+            let originFileName = path.substring(path.indexOf("_") + 1);
+
+            const $div = document.createElement('div');
+            $div.classList.add('thumbnail-box');
+
+
+            //이미지인지 확장자 체크
+            if (isImageFile(originFileName)) {
+               //이미지인 경우
+               originFileName = originFileName.substring(originFileName.indexOf("_") + 1);
+
+               const $img = document.createElement('img');
+               $img.setAttribute('src', '/loadFile?fileName=' + path);
+               $img.setAttribute('alt', originFileName);
+
+               $div.appendChild($img);
+
+
+            } else {
+               //이미지가 아닌 경우: 다운로드 링크 생성
+               const $link = document.createElement('a');
+               $link.setAttribute('href', '/loadFile?fileName=' + path);
+
+               $link.innerHTML = '<img src="/img/file_icon.jpg" alt="파일아이콘"> <span class="file-name">' +
+                  originFileName + '</span>';
+
+               $div.appendChild($link);
+            }
+            $attachDiv.append($div);
+         }
+
+         //정규표현식으로 이미지파일 여부 확인하는 함수
+         function isImageFile(originFileName) {
+            const pattern = /jpg$|gif$|png$/i;
+            return originFileName.match(pattern);
+         }
+
+      }); //end jquery
+   </script>
+
+
+   <!-- 댓글 관련 스크립트 -->
    <script>
       // 댓글 처리 JS
       $(function () {
          //원본글 번호
          const boardNo = '${article.boardNo}';
+
          //날짜 포맷 변환 함수
          function formatDate(datetime) {
             //문자열 날짜 데이터를 날짜객체로 변환
@@ -143,6 +239,7 @@
             let day = dateObj.getDate();
             let hour = dateObj.getHours();
             let minute = dateObj.getMinutes();
+
             //오전, 오후 시간체크
             let ampm = '';
             if (hour < 12 && hour >= 6) {
@@ -158,40 +255,51 @@
             } else {
                ampm = '새벽';
             }
+
             //숫자가 1자리일 경우 2자리로 변환
             (month < 10) ? month = '0' + month: month;
             (day < 10) ? day = '0' + day: day;
             (hour < 10) ? hour = '0' + hour: hour;
             (minute < 10) ? minute = '0' + minute: minute;
+
             return year + "-" + month + "-" + day + " " + ampm + " " + hour + ":" + minute;
+
          }
+
          //댓글 페이지 태그 생성 배치함수
          function makePageInfo(pageInfo) {
             let tag = "";
+
             const begin = pageInfo.beginPage;
             const end = pageInfo.endPage;
+
             //이전 버튼 만들기
             if (pageInfo.prev) {
                tag += "<li class='page-item'><a class='page-link page-active' href='" + (begin - 1) +
                   "'>이전</a></li>";
             }
+
             //페이지 번호 리스트 만들기
             for (let i = begin; i <= end; i++) {
                const active = (pageInfo.criteria.page === i) ? 'page-active' : '';
                tag += "<li class='page-item'><a class='page-link page-custom " + active + "' href='" + i + "'>" +
                   i + "</a></li>";
             }
+
             //다음 버튼 만들기
             if (pageInfo.next) {
                tag += "<li class='page-item'><a class='page-link page-active' href='" + (end + 1) +
                   "'>다음</a></li>";
             }
+
             //태그 삽입하기
             $(".pagination").html(tag);
          }
+
          //댓글 태그 생성, 배치 함수
          function makeReplyListDOM(replyMap) {
             let tag = '';
+
             for (let reply of replyMap.replyList) {
                tag += "<div id='replyContent' class='card-body' data-replyId='" + reply.replyNo + "'>" +
                   "    <div class='row user-block'>" +
@@ -210,12 +318,15 @@
                   "    </div>" +
                   " </div>";
             }
+
             //만든 태그를 댓글목록 안에 배치
             $('#replyData').html(tag);
             //댓글 수 배치
             $('#replyCnt').text(replyMap.count);
+
             //페이지 태그 배치
             makePageInfo(replyMap.pageInfo);
+
          }
 
          //댓글 목록 비동기 요청처리 함수
@@ -227,8 +338,10 @@
                   makeReplyListDOM(replyMap);
                });
          }
+
          //페이지 첫 진입시 비동기로 댓글목록 불러오기
          getReplyList(1);
+
          //페이지 버튼 클릭 이벤트
          $('.pagination').on('click', 'li a', e => {
             e.preventDefault();
@@ -237,13 +350,14 @@
 
          //댓글 등록 버튼 클릭 이벤트
          $('#replyAddBtn').on('click', e => {
-            //서버로 댓글 내용을 전송해서 db에 저정
+
+            //서버로 댓글 내용을 전송해서 DB에 저장
             const reqInfo = {
                method: 'POST', //요청 방식
                headers: { //요청 헤더 내용
                   'content-type': 'application/json'
                },
-               //서버로 전송할 테이터
+               //서버로 전송할 데이터 (JSON)
                body: JSON.stringify({
                   boardNo: boardNo,
                   replyText: $('#newReplyText').val(),
@@ -255,42 +369,47 @@
                .then(msg => {
                   if (msg === 'insertSuccess') {
                      getReplyList(1);
-                     $('#replyText').val('');
-                     $('#replyWriter').val('');
+                     $('#newReplyText').val('');
+                     $('#newReplyWriter').val('');
                   } else {
-                     alert('댓글 등록에 실패');
+                     alert('댓글 등록에 실패했습니다.');
                   }
                })
          });
 
-         //댓긓 수정버튼 클릭 이벤트
+         //댓글 수정버튼 클릭 이벤트
          const $modal = $('#replyModifyModal');
          $('#replyData').on('click', '#replyModBtn', e => {
-            console.log('수정 버튼');
+            //console.log("수정버튼 클릭!");
             //모달 띄우기
             $modal.modal('show');
 
-            //기존 댓글 
+            //기존 댓글내용 가져오기
             const originText = e.target.parentNode.previousElementSibling.textContent;
             // console.log(originText);
+
             $('#modReplyText').val(originText);
 
             //모달이 열릴때 모달안에 댓글번호 넣어놓기
             const replyId = e.target.parentNode.parentNode.parentNode.dataset.replyid;
+            // console.log(replyId);
+
             $('#modReplyId').val(replyId);
-            // console.log(replyid);
          });
-         //모달창 닫기
-         $('.modal-header button').on('click', e => {
+
+         //모달창 닫기 이벤트
+         $('.modal-header button, .modal-footer button:last-child').on('click', e => {
             $modal.modal('hide');
          });
 
          //댓글 수정 요청 이벤트
          $('#replyModBtn').on('click', e => {
-            // 댓글 번호
+            //댓글 번호
             const replyId = $('#modReplyId').val();
-            //댓긇 내용
+            //수정된 댓글 내용
             const replyText = $('#modReplyText').val();
+            //console.log("댓글번호:", replyId);
+            //console.log("댓글내용:", replyText);
 
             const reqInfo = {
                method: 'PUT',
@@ -301,7 +420,6 @@
                   replyNo: replyId,
                   replyText: replyText
                })
-
             };
             fetch('/api/v1/reply/' + replyId, reqInfo)
                .then(res => res.text())
@@ -310,18 +428,18 @@
                      $modal.modal('hide');
                      getReplyList(1);
                   } else {
-                     alert("댓글 수정에 실패");
+                     alert("댓글 수정에 실패했습니다.");
                   }
                })
          });
 
          //댓글 삭제 비동기 요청 이벤트
-         $('#replyData').on('click', '#replyDelBtn', e => {
-            //댓글 번호
+         $("#replyData").on("click", "#replyDelBtn", e => {
             const replyId = e.target.parentNode.parentNode.parentNode.dataset.replyid;
-
-            console.log(replyId);
-
+            //console.log("삭제 버튼 클릭! : " + replyId);
+            if (!confirm("진짜로 삭제할거니??")) {
+               return;
+            }
             const reqInfo = {
                method: 'DELETE'
             };
@@ -329,16 +447,15 @@
                .then(res => res.text())
                .then(msg => {
                   if (msg === 'delSuccess') {
-                     $modal.modal('hide');
                      getReplyList(1);
                   } else {
-                     alert("삭제 실패");
+                     alert("댓글 삭제에 실패했습니다.");
                   }
                })
          });
-
       });
    </script>
+
 
 </body>
